feat(strategy): add HTF bias score system (v2.35.0) and fix spread calculation
Replace binary HTF alignment filter with continuous bias score (0.0-1.0)
using professional algo trader approach. Add per-pair modes (active/monitor/
disabled) and confidence multiplier based on bias score.

Also includes:
- Fix MINI epic scalingFactor inference for spread calculation
- Fix CEEM price normalization in trade sync
- Correct scalp HTF timeframe docs (1h default, not 15m)
- Add percentage displays to rejection analysis charts
- Add date sorting options to signals page
- Use scalp-specific SL/TP in order executor

+++ dev-app/routers/orders_router.py
@@ -52,8 +52,8 @@ class TradeRequest(BaseModel):
 
     # Optional fields
     size: Optional[float] = None
-    stop_distance: Optional[int] = None
-    limit_distance: Optional[int] = None
+    stop_distance: Optional[float] = None
+    limit_distance: Optional[float] = None
     use_provided_sl_tp: Optional[bool] = False  # Use scanner-provided SL/TP values
     custom_label: Optional[str] = None
     risk_reward: Optional[float] = 2.0  # Default RR
